refactor(phase5): extract LoadingScreen, KanbanModal, useDeadlineAlert, TodoDeadlinePanel, TodoMetaBadges
ST5-1: LoadingScreen extracted from comm-time.tsx (107-line JSX → 1-line import)
ST5-2: KanbanModal extracted; absorbs showStatusManager state from CommTimeComponent
ST5-3: useDeadlineAlert hook extracted; minute-debounce + threshold algorithm verbatim
ST5-4: TodoDeadlinePanel + TodoMetaBadges extracted from TodoListPanel (938→~790 lines)

All 442 tests pass, TypeScript clean.
